requirements.txt
anyio==3.5.0
asgiref==3.5.1
autopep8==1.6.0
cffi==1.15.1
click==8.1.3
colorama==0.4.4
cryptography==37.0.4
fastapi==0.77.1
h11==0.13.0
httptools==0.4.0
idna==3.3
mysql-connector==2.2.9
pycodestyle==2.8.0
pycparser==2.21
pydantic==1.9.0
python-dateutil==2.8.2
python-dotenv==0.20.0
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
typing-extensions==4.2.0
uvicorn==0.17.6
watchgod==0.8.2
websockets==10.3
